% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{actSim}
\alias{actSim}
\title{Average commute time}
\usage{
actSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
This function calculates vertex similarity based on the average number of
steps, that random walker on the graph needs to get from one vertex to
another.
}

