% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{bcloseCentrality}
\alias{bcloseCentrality}
\title{Closeness centrality for bipartite graphs.}
\usage{
bcloseCentrality(g)
}
\arguments{
\item{g:}{igraph object}
}
\value{
Returns a list scores of closeness centrality of the nodes.
}
\description{
Measures Closeness centrality of bipartite graphs . TThe closeness centrality of a vertex is inversely proportional to the total geodesic distance to all other vertices.
This function is the adaption to bipartite graphs as presented in Borgatti and Everett (1997).
}
\details{
Closeness centrality for bipartite graphs
}
\references{
\itemize{
  \item Borgatti, S. P. and Everett, M. G. (1997) Network analysis of 2--mode data. Social Networks \bold{19}, 243--269.
  }
}

