% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{bi.BetweenessCentrality}
\alias{bi.BetweenessCentrality}
\title{Betweeness centrality for bipartite graphs.}
\usage{
bi.BetweenessCentrality(g, SM = TRUE)
}
\arguments{
\item{g:}{igraph object}

\item{SM:}{TRUE or FALSE}
}
\value{
Returns a list scores of closeness centrality of the nodes. By deafult \code{SM=TRUE} which return the betweeness centrality for two different sets of nodes.
}
\description{
Measures Betweeness centrality of bipartite graphs. The betweenness centrality of a vertex  may be roughly defined as the number of geodesic paths that pass through a given vertex,
weighted inversely by the total number of equivalent paths between the same two vertices, including those that do not pass through the given vertex.
It also has way to calculate single mode betweeness centrality. Single mode centralization for bipartite graphs measures the extent to which vertices in one vertex subset are central
relative only to other vertices in the same subset.This function is the adaption to bipartite graphs as presented in Borgatti and Everett (1997).
}
\details{
Betweeness centrality for bipartite graphs
}
\references{
\itemize{
  \item Borgatti, S. P. and Everett, M. G. (1997) Network analysis of 2--mode data. Social Networks \bold{19}, 243--269.
  }
}

