% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{bi.degreeCentrality}
\alias{bi.degreeCentrality}
\title{Degree centrality for Bipartite Graphs.}
\usage{
bi.degreeCentrality(g, loops = FALSE, SM = TRUE)
}
\arguments{
\item{g:}{igraph object}
}
\value{
Returns a list scores of the degree centrality the nodes in the rows are scored first followed by column vertices.
By deafult \code{SM=TRUE} which return the betweeness centrality for two different sets of nodes.
}
\description{
Measures graph degree centrality for bipartite graphs as well as Unipartite Graphs.The degree centrality of a vertex can be defined as fraction of incident edges over the number of all possible edges.
This function is the adaption to bipartite graphs as presented in Borgatti and Everett (1997).
}
\details{
Compute Degree Centrality of the Bipartite Graph
}
\references{
\itemize{
  \item Borgatti, S. P. and Everett, M. G. (1997) Network analysis of 2--mode data. Social Networks \bold{19}, 243--269.
  }
}

