% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{bi.Density}
\alias{bi.Density}
\title{Graph Density for Bipartite graphs.}
\usage{
bi.Density(g)
}
\arguments{
\item{g:}{igraph object}
}
\value{
Returns a list scores of the degree Density the nodes in the rows are scored first followed by column vertices.
}
\description{
Measures graph density  of bipartite graphs . The density captures the fraction of actual present edges over the number of all possible edges, given that multiple edges are not allowed.
This function is the adaption to bipartite graphs as presented in Borgatti and Everett (1997).
}
\details{
Compute Graph density Bipartite Graph
}
\references{
\itemize{
  \item Borgatti, S. P. and Everett, M. G. (1997) Network analysis of 2--mode data. Social Networks \bold{19}, 243--269.
  }
}

