% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{bi.weightedProjection}
\alias{bi.weightedProjection}
\title{Projecting Bipartite Networks.}
\usage{
bi.weightedProjection(g, vertex = FALSE, mode = "shared-neighbours",
  weight = FALSE)
}
\arguments{
\item{g:}{igraph object}

\item{mode:}{\code{'shared-neighbours'} and \code{'newman'}}
}
\value{
Returns a igraph object of the projected matrix with edge weights.
}
\description{
Projecting the Bipartite Network based on the vertex type (\code{TRUE} or \code{FALSE}).The mode \code{"shared-neighbours"} adds the number of shared neighbours in the bipartite graph to the edge linking two neighbours in the monopartite projection.
\code{"newman"} adopts the weighting scheme presented by Newman (2001), that weighs the contributions of shared neighbours by the size of their linkage profiles minus one.
}
\details{
Projecting Bipartite Networks
}
\references{
\itemize{
  \item Newman M (2001) Scientific collaboration networks. II. Shortest paths, weighted networks, and centrality. Physical Review E 64.
  }
}

