% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetWalk.R
\name{biNetwalk}
\alias{biNetwalk}
\title{Bipartite Random Walk}
\usage{
biNetwalk(g1, s1, s2, normalise = c("laplace", "none"), setSeeds = NULL,
  file = NULL, restart = 0.8, parallel = TRUE, multicores = NULL,
  verbose = T, weight = FALSE)
}
\arguments{
\item{g1:}{igraph object}

\item{s1:}{Accepts a matrix object of similarity scores for targets.}

\item{s2:}{Accepts a matrix object similarity scores for compounds.}

\item{normalise:}{Normalisation of matrix using laplacian or None(the transition matrix will be column normalized)}

\item{setSeeds:}{seeds file}

\item{file:}{Accepts file for seeds}

\item{restart:}{restart value}

\item{parallel:}{parallel performance either True or False . Parallelization is implemented using foreach.}

\item{multicore:}{using multicores}
}
\description{
Randomm walk with restart on Bipartite networks
}
\references{
\itemize {
\item Chen X, et al. Drug–target interaction prediction by random walk on the heterogeneous network. Mol. BioSyst 2012;8:1970-1978.
\item Vanunu O, Sharan R. Proceedings of the German Conference on Bioinformatics. Germany: GI; 2008. A propagation-based algorithm for inferring gene-disease assocations; pp. 54–63.
}
}

