% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{cnSim}
\alias{cnSim}
\title{Common neighbors vertex similarity}
\usage{
cnSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
Similarity measure counting number of common neighbors.
}

