% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{cosineSim}
\alias{cosineSim}
\title{Cosine vertex similarity/ Salton index}
\usage{
cosineSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
This function measures the cosine of the angle between columns of
the adjacency matrix, corresponding to given nodes.
}

