% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{diceSim}
\alias{diceSim}
\title{Dice similarity}
\usage{
diceSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
Similarity measure based on dice similarity. This function measures a relative size of an intersection of neighbors sets
}

