% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{distSim}
\alias{distSim}
\title{Geodesic distance vertex similarity}
\usage{
distSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
This function calculates similarity score for vertices based on the shortest paths between them.
}

