% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{get.biWeightedProjection}
\alias{get.biWeightedProjection}
\title{Projecting Bipartite Networks.}
\usage{
get.biWeightedProjection(A,vertex=FALSE,mode='shared-neighbours')
}
\arguments{
\item{A}{Biparite matrix object}

\item{vertex}{Either TRUE of FALSE where if FALSE then off-diagonal elements represent edge weights, diagonal elements vertex weights}

\item{mode}{\code{'shared-neighbours'} and \code{'newman'}}
}
\value{
Returns a igraph object of the projected matrix with edge weights.
}
\description{
Projecting the Bipartite Network based on the vertex type (\code{TRUE} or \code{FALSE}).The mode \code{"shared-neighbours"} adds the number of shared neighbours in the bipartite graph to the edge linking two neighbours in the monopartite projection.
\code{"newman"} adopts the weighting scheme presented by Newman (2001), that weighs the contributions of shared neighbours by the size of their linkage profiles minus one.
}
\details{
Projecting Bipartite Networks
}
\references{
\itemize{
  \item Newman M (2001) Scientific collaboration networks. II. Shortest paths, weighted networks, and centrality. Physical Review E 64.
  }
}

