% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{get.Communities}
\alias{get.Communities}
\title{Get communities from a graph}
\usage{
get.Communities(g, num.nodes = 3, calgo = walktrap.community)
}
\arguments{
\item{g:}{igraph object}

\item{num.nodes:}{Number of nodes to keep in the community.}

\item{calgo:}{The community algorithm to use to find the communities.}
}
\description{
Get  the communities in a graph. It First decomposes the graph into components and then uses igraph's
various community detection algorithms to detect communities in a graph.
}
\details{
Get the communities in a graph
}

