% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getComm}
\alias{getComm}
\title{get the communities from the given network with communuty prediction algorithm}
\usage{
getComm(g, num.nodes = 3, calgo = walktrap.community)
}
\arguments{
\item{g:}{igraph object}

\item{num.nodes:}{Number of nodes to keep in the community.}

\item{calgo:}{The community algorithm to use to find the communities.}
}
\description{
Peforms random walk with restart with preferred seed sets.
}
\details{
Random walk on unipartite networks
}
\references{
\itemize{
  \item Kohler S, et al. Walking the Interactome for Prioritization of Candidate Disease Genes. American Journal of Human Genetics. 2008;82:949–958.
  \item Can, T., Çamoǧlu, O., and Singh, A.K. (2005). Analysis of protein-protein interaction networks using random walks. In BIOKDD '05: Proceedings of the 5th international workshop on Bioinformatics (New York, USA: Association for Computing Machinery). 61–68
}
}

