% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{hdiSim}
\alias{hdiSim}
\title{Hub Depressed Index}
\usage{
hdiSim(graph)
}
\description{
This measures assigns lower scores to links adjacent to hubs (high degree nodes). It counts common neighbors of two vertices and weigths the result
}

