% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NetWalk.R
\name{hgviNet}
\alias{hgviNet}
\title{Heterogenous Graph based Inference on Bipartite Network}
\usage{
hgviNet(g1, s1, s2, alpha = 0.4, steps = 20)
}
\arguments{
\item{g1}{: igraph object}

\item{alpha}{: decay factor initialized to 0.4.}

\item{steps}{: Number of steps to iterate.}

\item{s1:}{Accepts a matrix object of similarity scores for targets.}

\item{s2:}{Accepts a matrix object similarity scores for compounds.}
}
\description{
Peforms computation of heterogenous graph inference explained by Wang etal.
}
\details{
Heterogenous Graph based Inference
}
\examples{
\donttest{
A <- enzyme_ADJ
S = enzyme_Csim
S1 = enzyme_Gsim
g1 = graph.incidence(A)
M1 <- hgviNet(A,s1=S,s2=S1,alpha=0.4,steps=20)
}
}
\references{
\itemize{
  \item Wang W, et al. Drug target predictions based on heterogeneous graph inference. Pac. Symp. Biocomput. 2013:53-64
}
#'
}

