% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{hpiSim}
\alias{hpiSim}
\title{Hub promoted Index}
\usage{
hpiSim(graph)
}
\description{
This measures assigns higher scores to links adjacent to hubs (high degree nodes). It counts common neighbors of two vertices and weigths the result
}

