% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{invLSim}
\alias{invLSim}
\title{Pseudoinverse of the Laplacian}
\usage{
invLSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
Similarity measure based solely on the pseudoinverse of the Laplacian matrix. This function counts all the paths between given pair of nodes, with shorter
}

