% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{jcSim}
\alias{jcSim}
\title{Jaccard Index similarity}
\usage{
jcSim(graph)
}
\description{
Similarity measure based on jaccard index.This is a simple wrapper to an \pkg{igraph} function \code{\link[igraph]{similarity.jaccard}},
}

