% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{katzSim}
\alias{katzSim}
\title{Katz Index similarity}
\usage{
katzSim(graph, beta = 0.001)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
Similarity measure based on all paths in a graph. This function counts all the paths between given pair of nodes, with shorter
}

