% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{lhnSim}
\alias{lhnSim}
\title{Leicht-Holme-Newman Index}
\usage{
lhnSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
This index gives high similarity to node pairs that have many common
neighbors compared not to the possible maximum, but to the expected number of such neighbors
}

