% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{lpSim}
\alias{lpSim}
\title{Local Paths Index}
\usage{
lpSim(graph, eps = 0.01)
}
\arguments{
\item{graph:}{igraph object}

\item{eps:}{epsilon parameter}
}
\value{
a matrix object of similarity scores.
}
\description{
This function counts the number of two-paths and three-paths between nodes,
}

