% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetWalk.R
\name{nbiNet}
\alias{nbiNet}
\title{Network Based Inference}
\usage{
nbiNet(A, alpha = 0.5, lamda = 0.5, s1 = NA, s2 = NA,
  format = c("igraph", "matrix", "pairs"))
}
\arguments{
\item{A:}{Adjacency Matrix}

\item{lamda:}{lamda parameter}

\item{alpha:}{alpha parameter}

\item{S1:}{Target Similarity matrix}

\item{S2:}{Chemical Similarity Matrix}

\item{format:}{type of file as Adjacnency file}
}
\description{
Given a bipartite graph , a two phase resource transfer Information from  X(x,y,z) set of nodes gets distributed to Y set of nodes and then again goes back to resource X.
This process allows us to define a technique for the calculation of the weight matrix W.
}
\details{
Network based inference on Bipartite networks
}
\examples{
\donttest{
A <- enzyme_ADJ
S1 = enzyme_Csim
S2 = enzyme_Gsim
g1 = graph.incidence(A)
## other format available \\code{format = c("igraph","matrix","pairs")}
M2 <- nbiNet(A,alpha=0.5, lamda=0.5,  S1=S1, S2=S2,format = "matrix")
}
}
\references{
\itemize{
\item Cheng F, et al. Prediction of drug-target interactions and drug repositioning via network-based inference. PLoS Comput. Biol. 2012;8:e1002503.
\item Zhou T, et al. Solving the apparent diversity-accuracy dilemma of recommender systems. Proc. Natl Acad. Sci. USA 2010;107:4511-4515.
\item Zhou T, et al. Bipartite network projection and personal recommendation. Phys. Rev. E Stat. Nonlin. Soft Matter Phys. 2007;76:046115.
\item Blog post from Abhik Seal \url{http://data2quest.blogspot.com/2015/02/link-prediction-using-network-based.html}
}
}

