% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netpredictor.R
\name{net.perf}
\alias{net.perf}
\title{Link Prediction Performance}
\usage{
net.perf(A, S1, S2, relinks = 100, numT = 2, Calgo = c("rwr", "nbi",
  "netcombo", "all"))
}
\arguments{
\item{A}{: Drug target association matrix}

\item{S1}{: Sequence similarity matrix object}

\item{S2:}{Accepts a matrix object similarity scores for compounds.}

\item{relinks:}{Number of links to remove randomly from the input matrix.}

\item{numT:}{Frequency of the number of targets.}

\item{Calgo:}{Algorithm to use for Bipartite link prediction options are "rwr","nbi" & "netcombo".}

\item{norm:}{normalization of matrices options are "laplace" or "none".}
}
\value{
it returns a list of aucc,auc, bedorc,enrichment factor and auc (top 10%)
\itemize{
  \item {Truchon et al. Evaluating Virtual Screening Methods: Good and Bad Metrics for the "Early Recognition" Problem. J. Chem. Inf. Model. (2007) 47, 488-508.}
  \item {Sheridan RP et al. Protocols for bridging the peptide to nonpeptide gap in topological similarity searches. J. Chem. Inf. Comput. Sci. (2001) 41, 1395-1406.}
}
}
\description{
This function samples links and removies links from the adjacency matrix and predicts them and calculates  area under accumulation curve, AUC, bedroc, and Enrichment factor.
}
\details{
get the performance of the link Prediction algorithms.
}
\examples{
\dontrun{
data(Enzyme)
A = enzyme_ADJ
S1 = enzyme_Gsim
S2= enzyme_Csim
m = net.perf(A,S1,S2,relinks = 50,numT=2,Calgo="nbi")
}
}

