% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetWalk.R
\name{netPredPerf}
\alias{getTopresults}
\alias{netPredPerf}
\title{Get Top Results}
\usage{
getTopresults(A, P, top = 10, druglist = NULL)
}
\arguments{
\item{A}{: Drug target association matrix.}

\item{P:}{Drug target predicted matrix.}

\item{top:}{top number of predicted targets.}

\item{druglist:}{It accepts a vector of drugnames for which results will return}
}
\value{
it returns a list of aucc,auc, bedorc,enrichment factor and auc (top 10%)
}
\description{
The function returns the given top number of predicted results along with true interactions.
}
\details{
Get top predicted results.
}
\examples{
\donttest {
data(Enzyme)
A = enzyme_ADJ
S1 = enzyme_Gsim
S2= enzyme_Csim
## Running the netcombo algorithm.
P <- netCombo(g1,s1=S1,s2=S2,nbi.alpha=0.5,nbi.lamda=0.5,par=TRUE)
result = getTopresults(A,P,top=10,druglist=NULL)
## Getting result from a drug list.
drugs = c("D00014","D00018", "D00029", "D00036","D00045","D00049")
result = getTopresults(A,P,top=10,druglist=drugs)
}
}

