% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{paSim}
\alias{paSim}
\title{Preferential attachment vertex similarity}
\usage{
paSim(graph)
}
\arguments{
\item{graph:}{a igraph object}
}
\description{
This function calculates similarity score for vertices based on their degrees.
}

