% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.Community}
\alias{plot.Community}
\title{plotting Communities}
\usage{
\method{plot}{Community}(gc, cols = 3)
}
\description{
This uses an object of getCommuntiy class and extracts parameters to plot current communities.
}

