% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{raSim}
\alias{raSim}
\title{Similarity measure based on resource allocation process.}
\usage{
raSim(graph)
}
\arguments{
\item{graph:}{igraph object}
}
\description{
This function counts the number of common neighbours weighted by the inverse
of their degrees.
}

