% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{rwr}
\alias{rwr}
\title{performing random walk with restart both in parallel and non-parallel way.
It takes the transition matrix W the initial matrix P0 matrix parameter for parallization
and the number of cores to run parallelisation on. Also a restart parameter is available as r
to get better results with different datasets one needs to tune restart parameter r.}
\usage{
rwr(W, P0matrix, par = FALSE, r = 0.7, multicores = multicores)
}
\description{
performing random walk with restart both in parallel and non-parallel way.
It takes the transition matrix W the initial matrix P0 matrix parameter for parallization
and the number of cores to run parallelisation on. Also a restart parameter is available as r
to get better results with different datasets one needs to tune restart parameter r.
}

