% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{saveAsGEXF}
\alias{saveAsGEXF}
\title{save GML}
\usage{
saveAsGEXF(g, filepath = "output.gexf")
}
\arguments{
\item{g:}{igraph object}

\item{filepath:}{output graph name}
}
\value{
Outputs a Gephi format gexf graph
}
\description{
Export an igraph object to gexf format graph file which can be in opened in Gephi.
}
\details{
Export file to GML
}

