% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{tMat}
\alias{tMat}
\title{Get the transition matrix for a Bipartite Graph. Input a sequence similarity matrix (s1),
chemical similarity matrix (s2) and drug target adjacency matrix (g1) where rows are
protein targets and columns as Drugs.}
\usage{
tMat(g1, s1, s2, normalise = "laplace")
}
\description{
Get the transition matrix for a Bipartite Graph. Input a sequence similarity matrix (s1),
chemical similarity matrix (s2) and drug target adjacency matrix (g1) where rows are
protein targets and columns as Drugs.
}

