% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unipartite.R
\name{unetSim}
\alias{unetSim}
\title{unetSim}
\usage{
unetSim(graph, method)
}
\arguments{
\item{graph}{an object of class \code{igraph}}

\item{method}{a method (single string) for calculating similarities , see details for methods.}
}
\value{
a matrix of similarity scores.
}
\description{
This function calculates vertex proximity (similarity) with selected method
and between selected vertices.
}
\details{
Get similarity using various link prediction methods

Following methods are available:
\code{aa} Adamic-Adar index
\code{act} average commute time
\code{cn} common neighbours
\code{jc} cosine similarity
\code{dice} cosine similarity on L+
\code{invL} graph distance
\code{hdi} Hub Depressed Index
\code{hpi} Hub Promoted Index
\code{dist} Jaccard coefficient
\code{katz} Katz index
\code{lhn} Leicht-Holme-Newman Index
\code{lp} Local Path Index
\code{cosine} Matrix Forest Index
\code{pa} preferential attachment
\code{ra} resource allocation index
}
\examples{
\donttest{
# net <- netSim(graph=proj_weighted,method="pa")
}
}
\references{
\itemize{
  \item T. Zhou, L. Lv, and Y.-C. Zhang.Predicting missing links via local information, The European Physical Journal B - Condensed Matter and Complex Systems, vol. 71, no. 4, pp. 623-630, Oct. 2009.
  }
}

