% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/results.R
\name{get.candidate.graph}
\alias{get.candidate.graph}
\title{get.candidata.graph}
\usage{
get.candidate.graph(file = NULL, affinity = NULL, top = 10,
  format = c("bipart", "unipart"))
}
\arguments{
\item{file:}{Input seeds file must be 2 column dataframe or matrix of seed nodes of a bipartite graph.}

\item{affinity:}{Computed affinity matrix from Bipartite network.}

\item{top:}{retrieving top number of results.}
}
\description{
get.candidate.graph
}

