% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results.R
\name{saveGML}
\alias{saveGML}
\title{save GML}
\usage{
saveGML(g, fileName, title = "untitled")
}
\arguments{
\item{g:}{igraph object}

\item{fileName:}{exported file name}

\item{title:}{title of the exported file name}
}
\description{
Export an igraph object to GML file which can be in Cytoscape or Gephi.
}
\details{
Export file to GML
}

