% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NetWalk.R
\name{Significant network}
\alias{Significant network}
\alias{sig.net}
\title{Significant Network}
\usage{
sig.net(data, g, Amatrix, num.permutation = 10, adjp.cutoff = 0.05,
  p.adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
  "fdr"), parallel = TRUE, multicores = NULL, verbose = T)
}
\arguments{
\item{data:}{n x m Adjancency matrix of seeds or dataframe of pairs.}

\item{g:}{igrah object of bipartite indcident adjacency matrix.}

\item{Amatrix:}{Affinity matrix computed from biNetWalk or uNetWalk.}

\item{num.permutation:}{number of permutation of Affinity matrix needed to performed.}

\item{adjp.cutoff:}{pvalue cutoff 0.05}

\item{p.adjust.method:}{Adjusting the pvalue by diiferent method.It uses method from the stats package.}

\item{parallel:}{Using parallization either True or False}

\item{multicores:}{If parallisation is set TRUE number of cores to perform parallel computaion.}

\item{Verbose:}{For verbose output.}
}
\description{
Significance of Bipartite networks
}
\examples{
\donttest{
data(Enzyme)
A <- enzyme_ADJ
S = enzyme_Csim
S1 = enzyme_Gsim
g1 = graph.incidence(A)
Q = biNetwalk(g1,S,S1,normalise="laplace",parallel=FALSE,verbose=T)
Z = sig.net(data=A,g=g1,Amatrix=Q,num.permutation=100,adjp.cutoff=0.01,p.adjust.method="BH",parallel=FALSE)
}
}

